feat: implement iterator structures for catalog management, adjust flight validation for phase2 and initialize project structure for phase2 (#62)

+++ trabalho-pratico/src/utils/date.c
@@ -3,7 +3,6 @@
 #include <stdlib.h>
 #include <string.h>
 
-
 int string_to_date(const char* str) {
     if (!str || strcmp(str, "N/A") == 0) return -1; ///< Considerar "N/A" como -1
 
@@ -23,7 +22,7 @@ long string_to_datetime(const char* str) {
     int h = (str[11]-'0')*10 + (str[12]-'0');
     int min = (str[14]-'0')*10 + (str[15]-'0');
     
-    return y*100000000L + m*1000000L + d*10000L + h*100 + min; ///< O 'L' força o cálculo a ser feito em long (evita overflow)
+    return (long)y*100000000L + m*1000000L + d*10000L + h*100 + min; ///< O 'L' força o cálculo a ser feito em long (evita overflow)
 }
 
 int get_date_part(long datetime) {
